fix: extract bdd page template to separate file to resolve unescaped backtick parsing errors
